PredmetIGrupaRepository finished

+++ app/src/main/java/ba/etf/rma21/projekat/data/models/Grupa.kt
@@ -3,7 +3,7 @@ package ba.etf.rma21.projekat.data.models
 import com.google.gson.annotations.SerializedName
 
 data class Grupa(
-    @SerializedName("id") val id : Long,
+    @SerializedName("id") val id : Int,
     @SerializedName("naziv") val naziv : String) {
     override operator fun equals(other: Any?): Boolean {
         if (other == null)
