+++ app/src/main/java/ba/etf/rma21/projekat/data/models/KvizTaken.kt
@@ -4,7 +4,7 @@ import com.google.gson.annotations.SerializedName
 import java.util.*
 
 data class KvizTaken(
-    @SerializedName("id") var id : Long,
+    @SerializedName("id") var id : Int,
     @SerializedName("student") var student : String,
     @SerializedName("datumRada") var datumRada : Date,
     @SerializedName("osvojeniBodovi") var osvojeniBodovi : Int
