+++ app/src/main/java/ba/etf/rma21/projekat/data/models/Odgovor.kt
@@ -3,7 +3,7 @@ package ba.etf.rma21.projekat.data.models
 import com.google.gson.annotations.SerializedName
 
 class Odgovor(
-    @SerializedName("id") var id : Long,
+    @SerializedName("id") var id : Int,
     @SerializedName("odgovoreno") var odgovoreno : Int
 ) {
