+++ app/src/main/java/ba/etf/rma21/projekat/data/models/Predmet.kt
@@ -3,7 +3,7 @@ package ba.etf.rma21.projekat.data.models
 import com.google.gson.annotations.SerializedName
 
 data class Predmet(
-    @SerializedName("id") val id : Long,
+    @SerializedName("id") val id : Int,
     @SerializedName("naziv") val naziv: String,
     @SerializedName("godina") val godina: Int) {
     override operator fun equals(other: Any?): Boolean {
